move ds files under ds folder
