+++ src/com/i2017/dataStructure/ListNode.java
@@ -1,4 +1,4 @@
-package com.i2017.algorithm;
+package com.i2017.dataStructure;
 
 public class ListNode {
 	private int value;
