+++ src/com/i2017/dataStructure/Test.java
@@ -1,4 +1,4 @@
-package com.i2017.algorithm;
+package com.i2017.dataStructure;
 
 public class Test {
 	public static void main(String args[]){
